Extract transpilation orchestration into dedicated module
Refactor watch command to improve separation of concerns by extracting
transpilation coordination logic into a new dedicated module. This reduces
the watch.ts file size from 1207 lines to 872 lines (335 line reduction),
bringing it under the 1000-line guideline.

Changes:
- Created new transpilation coordinator module (src/cli/src/modules/transpilation/)
- Extracted transpileFile() function with complete lifecycle management
- Extracted displayTranspilationStatistics() function
- Moved TranspilationMetrics, TranspilationError, and TranspilationContext types
- Updated watch command to use the new coordinator
- All tests pass (346 tests)
- No lint errors
